encore: Adds 'project template' projects.
This adds a template for an empty Ext.NET Core project. Once built, it
can be installed with:

- dotnet new --install enlib.tpl.Empty

Then a new project can be created using the enlib project with:

- dotnet new extnet

This also includes:

- vscode build tasks
- changes to Visual Studio solution file
- common/template.targets file with project rules that should be
  included on any 'project template' project.

The projects in Templates/ are actually simple projects set up to use
the enlib NuGet package. They can be built and run like any project, and
like the final template project would.

The projects in Templates/Package are special projects just to compile
the template projects into the respective NuGet package, with specific
rules to make the NuGet package a template package.